Split temperature reading around wifi connect
When reading the temperatures, it's necessary to wait for 750ms after
issuing the convert_temp call before reading the results. Splitting the
process into 2 functions means that we can do the wifi connection while
we're waiting for the temperature read to happen, which reduces the
overall time that the device needs to be awake, which should reduce the
power consumption.